feat: [ft-wave1-inference-selection-004] - Close out ownership metadata and verification gates
