+++ docs/temp/functional-tests-expansion/migration-ledger.md
@@ -510,6 +510,9 @@ batch ids (FND-007-003…005). Remaining non-catch-all packages use `n/a` for
 | tests/functional/workers/inference/failure_normalization_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderNonZeroExitMapsToPublicFailure | short | tests/functional/workers/inference/failure_normalization_test.go | none | none | n/a |
 | tests/functional/workers/inference/failure_normalization_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderAuthRateLimitAndTimeoutRemainDistinct | short | tests/functional/workers/inference/failure_normalization_test.go | none | none | n/a |
 | tests/functional/workers/inference/failure_normalization_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderFailureRedactsPromptEnvironmentAndCredentials | short | tests/functional/workers/inference/failure_normalization_test.go | none | none | n/a |
+| tests/functional/workers/inference/selection_test.go | you-agent-factory/tests/functional/workers/inference | TestExplicitProviderAndModelReachSelectedProviderEdge | short | tests/functional/workers/inference/selection_test.go | none | none | n/a |
+| tests/functional/workers/inference/selection_test.go | you-agent-factory/tests/functional/workers/inference | TestWorkerProviderOverridesGlobalDefault | short | tests/functional/workers/inference/selection_test.go | none | none | n/a |
+| tests/functional/workers/inference/selection_test.go | you-agent-factory/tests/functional/workers/inference | TestUnknownProviderFailsBeforeProcessStart | short | tests/functional/workers/inference/selection_test.go | none | none | n/a |
 | tests/functional/workers/inference/process_cleanup_long_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderCancellationTerminatesCompanionProcesses | functionallong | tests/functional/workers/inference/process_cleanup_test.go | none | script-timeout-companion-smoke-100 | n/a |
 | tests/functional/workers/inference/process_cleanup_test.go | you-agent-factory/tests/functional/workers/inference | TestProcessTreeHelper | short | tests/functional/workers/inference/process_cleanup_test.go | none | none | n/a |
 | tests/functional/workers/inference/process_cleanup_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderSuccessWaitsForProcessAndStreamClosure | short | tests/functional/workers/inference/process_cleanup_test.go | none | none | n/a |
@@ -905,6 +908,9 @@ Batch execution guidance for later move work:
 | tests/functional/workers/inference/failure_normalization_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderNonZeroExitMapsToPublicFailure | short | tests/functional/workers/inference/failure_normalization_test.go | none | none | n/a |
 | tests/functional/workers/inference/failure_normalization_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderAuthRateLimitAndTimeoutRemainDistinct | short | tests/functional/workers/inference/failure_normalization_test.go | none | none | n/a |
 | tests/functional/workers/inference/failure_normalization_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderFailureRedactsPromptEnvironmentAndCredentials | short | tests/functional/workers/inference/failure_normalization_test.go | none | none | n/a |
+| tests/functional/workers/inference/selection_test.go | you-agent-factory/tests/functional/workers/inference | TestExplicitProviderAndModelReachSelectedProviderEdge | short | tests/functional/workers/inference/selection_test.go | none | none | n/a |
+| tests/functional/workers/inference/selection_test.go | you-agent-factory/tests/functional/workers/inference | TestWorkerProviderOverridesGlobalDefault | short | tests/functional/workers/inference/selection_test.go | none | none | n/a |
+| tests/functional/workers/inference/selection_test.go | you-agent-factory/tests/functional/workers/inference | TestUnknownProviderFailsBeforeProcessStart | short | tests/functional/workers/inference/selection_test.go | none | none | n/a |
 | tests/functional/workers/inference/process_cleanup_long_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderCancellationTerminatesCompanionProcesses | functionallong | tests/functional/workers/inference/process_cleanup_test.go | none | script-timeout-companion-smoke-100 | n/a |
 | tests/functional/workers/inference/process_cleanup_test.go | you-agent-factory/tests/functional/workers/inference | TestProcessTreeHelper | short | tests/functional/workers/inference/process_cleanup_test.go | none | none | n/a |
 | tests/functional/workers/inference/process_cleanup_test.go | you-agent-factory/tests/functional/workers/inference | TestProviderSuccessWaitsForProcessAndStreamClosure | short | tests/functional/workers/inference/process_cleanup_test.go | none | none | n/a |
